# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Item:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  request.CreateItemRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  request.UpdateItemRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  service.PaginatedItems:
    properties:
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      totalCount:
        type: integer
      totalPages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample Go API application with JWT Authentication, SQLC,
    and PostgreSQL.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go API Application
  version: "1.0"
paths:
  /admin/items:
    post:
      consumes:
      - application/json
      description: Creates a new item with a name and description. Requires JWT authentication
        and 'admin' role.
      parameters:
      - description: Item creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created item
          schema:
            $ref: '#/definitions/model.Item'
        "400":
          description: 'message: Invalid request data'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'message: Authentication token required / Invalid token'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'message: You do not have permission to access this resource.'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'message: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new item
      tags:
      - items
  /admin/items/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an item by its ID. Requires JWT authentication and 'admin'
        role.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: 'message: Invalid item ID format'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'message: Authentication token required / Invalid token'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'message: You do not have permission to access this resource.'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'message: Item not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'message: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete an item
      tags:
      - items
    put:
      consumes:
      - application/json
      description: Updates an existing item's name and description. Requires JWT authentication
        and 'admin' role.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated item
          schema:
            $ref: '#/definitions/model.Item'
        "400":
          description: 'message: Invalid request data / Invalid item ID format'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'message: Authentication token required / Invalid token'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'message: You do not have permission to access this resource.'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'message: Item not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'message: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an existing item
      tags:
      - items
...
  /items:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of sample items, requires JWT authentication.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default 10)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of items
          schema:
            $ref: '#/definitions/service.PaginatedItems'
        "400":
          description: 'message: Invalid pagination parameters'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'message: Authentication token required / Invalid token'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'message: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get list of items
      tags:
      - items
  /items/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a single item by its ID. Requires JWT authentication.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Item details
          schema:
            $ref: '#/definitions/model.Item'
        "400":
          description: 'message: Invalid item ID format'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'message: Authentication token required / Invalid token'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'message: Item not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'message: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get item by ID
      tags:
      - items
  /items/search:
    get:
      consumes:
      - application/json
      description: Searches for items across multiple fields using Elasticsearch.
        Requires JWT authentication.
      parameters:
      - description: Search query string
        in: query
        name: q
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default 10)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of search results
          schema:
            $ref: '#/definitions/service.PaginatedItems'
        "400":
          description: 'message: Invalid parameters'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'message: Authentication token required / Invalid token'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'message: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Search items
      tags:
      - items
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"